[IMP] Preparing migration from 7.0 to 8.0 crm_claim_ext, test_flake8 and test_server
