+++ crm_claim_ext/wizard/get_empty_serial_view.xml
