+++ crm_claim_ext/wizard/picking_from_exchange_lines_wizard_view.xml
