+++ crm_claim_ext/wizard/picking_from_returned_lines_wizard_view.xml
